+++ src/libtrx/game/catalog.c
@@ -3,6 +3,7 @@
 #include "filesystem.h"
 #include "game/music/ids.h"
 #include "game/objects/ids.h"
+#include "game/sound/ids.h"
 #include "log.h"
 #include "memory.h"
 
@@ -24,6 +25,10 @@ static const M_ENTRY m_CatalogEntryDefs[] = {
 #define X_CATALOG_ID(uuid_str, enum_value)                                     \
     { CATALOG_OBJECTS, enum_value, uuid_str },
 #include "game/catalog_objects.def"
+#undef X_CATALOG_ID
+#define X_CATALOG_ID(uuid_str, enum_value)                                     \
+    { CATALOG_SAMPLES, enum_value, uuid_str },
+#include "game/catalog_samples.def"
 #undef X_CATALOG_ID
 };