+++ src/libtrx/game/shell/flow.c
@@ -152,6 +152,7 @@ const SHELL_ARGS *Shell_CommonInit(const SHELL_ARGS *const args)
 
     M_LoadCatalog(CATALOG_OBJECTS, "catalog_objects.csv");
     M_LoadCatalog(CATALOG_MUSIC, "catalog_music.csv");
+    M_LoadCatalog(CATALOG_SAMPLES, "catalog_samples.csv");
 
     if (args->test_replay_path != nullptr) {
         SHELL_ARGS *tmp_args = TestReplay_Open(args->test_replay_path);