+++ docs/tr1/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Unreleased](https://github.com/LostArtefacts/TRX/compare/tr1-4.15.1...develop) - ××××-××-××
 - added a game flow option for cold water in custom levels, similar to TR3 (#4021)
 - added a splash effect when Lara jumps in wading depth water, similar to TR3+ (#3975)
-- added support for object ID and music ID catalog overrides
+- added support for object ID, music ID and sound ID catalog overrides
     This lets the builders to choose which slots to use to provide the game with assets the game uses.
     In the future, this feature can be used to get "hard-coded" TR2 sounds to work in TR1, and vice versa.
     See [the documentation](../07-CATALOGS.md) for details.